Only change buffer type for strings

+++ src/hardware.jl
@@ -81,7 +81,7 @@ on floor 1 and higher.
 """
 struct Location
     string::Int32
-    floor::Int32
+    floor::Int8
 end
 Base.isless(lhs::Location, rhs::Location) = lhs.string == rhs.string ? lhs.floor < rhs.floor : lhs.string < rhs.string
 
@@ -208,7 +208,7 @@ function read(filename::AbstractString, T::Type{Hydrophone})
             continue
         end
         string, floor, x, y, z = split(line)
-        location = Location(parse(Int32, string), parse(Int32, floor))
+        location = Location(parse(Int32, string), parse(Int8, floor))
         pos = Position(parse.(Float64, [x, y, z])...)
         push!(hydrophones, T(location, pos))
     end
